pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.9.1",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "flask-dance>=7.1.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn[gthread]>=23.0.0",
    "numpy>=2.3.2",
    "oauthlib>=3.3.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
    "flask-login>=0.6.3",
    "twilio>=9.8.0",
    "elevenlabs>=2.14.0",
    "pydub>=0.25.1",
    "requests-toolbelt>=1.0.0",
    "pytz>=2025.2",
]
